I is for Interface
Adding some not implemented methods,
Enum name cleanup
Prevent log event from modifying or cancelling logging

+++ Chraft/Plugins/Events/Args/EntityEventArgs.cs
@@ -68,10 +68,10 @@ namespace Chraft.Plugins.Events.Args
     }
     public class EntityAttackEventArgs : EntityEventArgs
     {
-        public int Damage { get; set; }
+        public short Damage { get; set; }
         public EntityBase EntityToAttack { get; set; }
 
-        public EntityAttackEventArgs(EntityBase entity, int damage, EntityBase entityToAttack)
+        public EntityAttackEventArgs(EntityBase entity, short damage, EntityBase entityToAttack)
             : base(entity)
         {
             Damage = damage;
